+++ Chirp/src/app/features/admin/components/admin-user-info/admin-user-info.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { User } from 'src/app/core/models/user.module';
+import { User } from 'src/app/core/models/user';
 
 @Component({
   selector: 'app-admin-user-info',
