+++ Chirp/src/app/shared/services/user.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { User } from 'src/app/core/models/user.module';
+import { User } from 'src/app/core/models/user';
 
 @Injectable({
   providedIn: 'root'
